+++ backend/src/main/java/eye/on/the/money/service/stock/RSUTaxService.java
@@ -156,8 +156,8 @@ public class RSUTaxService implements ICSVService {
                 .build();
     }
 
-    private BigDecimal value(BigDecimal price, Double quantity) {
+    private BigDecimal value(BigDecimal price, BigDecimal quantity) {
         if (price == null || quantity == null) return BigDecimal.ZERO;
-        return price.multiply(BigDecimal.valueOf(quantity)).setScale(2, RoundingMode.HALF_UP);
+        return price.multiply(quantity).setScale(2, RoundingMode.HALF_UP);
     }
 }
